Probe for the date of the last item
